move some actions to zauction specific folder
