+++ src/actions/index.ts
@@ -1,15 +1,14 @@
 export * from "./getDomainEvents";
 export * from "./getDomainMetadata";
 export * from "./getMetadataFromUri";
-export * from "./getPaymentTokenInfo";
 export * from "./lockDomainMetadata";
 export * from "./mintSubdomain";
 export * from "./setDomainMetadata";
 export * from "./setDomainMetadataUri";
 export * from "./setAndLockDomainMetadata";
 export * from "./setAndLockDomainMetadataUri";
-export * from "./setPaymentTokenForDomain";
 export * from "./setDomainRoyalty";
 export * from "./transferOwnership";
-export * from "./getZAuctionSpendAllowance";
-export * from './generateDefaultMetadata';
+export * from "./zauction/getZAuctionSpendAllowance";
+export * from "./generateDefaultMetadata";
+export * from "./zauction";
